# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1-100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by product menu category",
            "schema": {
              "$ref": "#/components/schemas/ProductCategory"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Full-text search across title and description",
            "schema": {
              "example": "latte",
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "minimum": 0,
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "minimum": 0,
              "example": 500,
              "type": "number"
            }
          },
          {
            "name": "available",
            "required": false,
            "in": "query",
            "description": "Filter to only available products",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "example": "newest",
              "type": "string",
              "enum": [
                "price_asc",
                "price_desc",
                "newest"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Search/list active products",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a single active product",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/admin/orders/{orderNumber}": {
      "get": {
        "operationId": "AdminOrdersController_getOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order details, including invoice information (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a delivery or pickup order (guest or authenticated)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_getOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get an order by number (owner, or guest with X-Order-Token)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{orderNumber}/screenshot": {
      "post": {
        "operationId": "OrdersController_uploadScreenshot",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Screenshot attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "summary": "Upload the payment screenshot for a pending order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{orderNumber}/cancel": {
      "patch": {
        "operationId": "OrdersController_cancelOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "summary": "Cancel a pending or confirmed order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/me/orders": {
      "get": {
        "operationId": "MyOrdersController_listMyOrders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1-100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated order history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user's order history",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/me/orders/{orderNumber}": {
      "get": {
        "operationId": "MyOrdersController_getMyOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of the current user's orders",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/products/{id}/reviews": {
      "get": {
        "operationId": "ProductReviewsController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1-100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated reviews",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List reviews for a product",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "BW Café Ordering API",
    "description": "Single-restaurant delivery/pickup ordering API — guest and registered checkout, admin order management.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "nullable": true,
            "additionalProperties": false,
            "example": null
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "ProductCategory": {
        "type": "string",
        "enum": [
          "coffee",
          "breakfast",
          "burger",
          "shawarma",
          "tacos",
          "drinks"
        ]
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "category": {
            "example": "coffee",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategory"
              }
            ]
          },
          "title": {
            "type": "string",
            "example": "Cappuccino"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Espresso with steamed milk foam"
          },
          "details": {
            "type": "string",
            "nullable": true,
            "example": "Contains dairy"
          },
          "price": {
            "type": "number",
            "example": 250
          },
          "quantity": {
            "type": "number",
            "nullable": true,
            "example": 50,
            "description": "null means stock is not tracked"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://res.cloudinary.com/demo/image/upload/bw-cafe/products/1.jpg"
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "category",
          "title",
          "price",
          "isAvailable",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductListDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "ProductListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "ProductDataDto": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductDto"
          }
        },
        "required": [
          "product"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "productId": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "productTitle": {
            "type": "string",
            "example": "Cappuccino"
          },
          "unitPrice": {
            "type": "number",
            "example": 250
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 500
          }
        },
        "required": [
          "id",
          "productTitle",
          "unitPrice",
          "quantity",
          "lineTotal"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "orderNumber": {
            "type": "string",
            "example": "BW-20260713-0042"
          },
          "customerName": {
            "type": "string",
            "example": "Customer Name"
          },
          "customerPhone": {
            "type": "string",
            "example": "+22200000000"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ],
            "example": "DELIVERY"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "Full delivery address as one string"
          },
          "pickupTime": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY",
              "OUT_FOR_DELIVERY",
              "COMPLETED",
              "CANCELLED",
              "REJECTED"
            ],
            "example": "PENDING"
          },
          "screenshotUrl": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "number",
            "example": 500
          },
          "deliveryFee": {
            "type": "number",
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 500
          },
          "currency": {
            "type": "string",
            "example": "MRU"
          },
          "customerNotes": {
            "type": "string",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-13T12:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "guestAccessToken": {
            "type": "string",
            "description": "Only present once, at creation, for guest orders"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "customerPhone",
          "orderType",
          "status",
          "subtotal",
          "deliveryFee",
          "total",
          "currency",
          "createdAt",
          "items"
        ]
      },
      "OrderListDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "OrderListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/OrderListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "OrderDataDto": {
        "type": "object",
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderDto"
          }
        },
        "required": [
          "order"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/OrderDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "minimum": 1,
            "example": 1
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 2
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150,
            "example": "Customer Name"
          },
          "customerPhone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "example": "+22200000000"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ],
            "example": "DELIVERY"
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "description": "Required for DELIVERY orders",
            "example": "Full delivery address as one string"
          },
          "pickupTime": {
            "type": "string",
            "description": "Required for PICKUP orders (ISO 8601)",
            "example": "2026-07-13T18:00:00.000Z"
          },
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          },
          "customerNotes": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "example": "Call before arriving"
          }
        },
        "required": [
          "customerName",
          "customerPhone",
          "orderType",
          "items"
        ]
      },
      "CancelOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "Changed my mind"
          }
        }
      },
...
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "productId": {
            "type": "number",
            "example": 3
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "nullable": true,
            "example": "Great coffee!"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-13T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "productId",
          "rating",
          "createdAt"
        ]
      },
...
      "ReviewListDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewDto"
            }
          },
          "total": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "ReviewListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ReviewListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
    }
  }
}